+++ src/features/useUncontrolledForm/components/Input/ui/Input.tsx
@@ -1,8 +1,8 @@
 import { FC, useState } from 'react';
-import { InputName, LabelTitle } from 'shared/types/InputTypes';
-import { InputWrap } from '../../InputWrap';
+import { InputName } from 'shared/types/InputTypes';
+import { InputWrap } from 'shared/ui/InputWrap';
 import { checkPassword } from 'shared/lib/handlingData';
-import './Input.css';
+import styles from './Input.module.css';
 
 interface InputProps {
   refs: React.MutableRefObject<HTMLInputElement>;
@@ -30,21 +30,19 @@ export const Input: FC<InputProps> = ({
         id={name}
         type={type}
         placeholder={title}
-        className={`inputForm ${isValid}`}
+        className={`${styles.inputForm} ${styles[`${isValid}`]}`}
         autoComplete="off"
         onChange={(e) => setValue(e.target.value)}
       />
-      {name === 'pass' &&
-        title !== LabelTitle.confirmPass &&
-        checkPassword(value) !== 1 && (
-          <progress
-            className="progressBar"
-            value={checkPassword(value)}
-            max={1}
-          />
-        )}
+      {name === 'pass' && checkPassword(value) !== 1 && (
+        <progress
+          className={styles.progressBar}
+          value={checkPassword(value)}
+          max={1}
+        />
+      )}
       {name === 'pass' && checkPassword(value) === 1 && (
-        <h3 className="passwordTitle">The password is strong</h3>
+        <h3 className={styles.passwordTitle}>The password is strong</h3>
       )}
     </InputWrap>
   );
