+++ src/features/useUncontrolledForm/ui/UncontrolledForm.module.css
@@ -11,25 +11,17 @@
   align-items: center;
 }
 
-.allInputs {
-  max-height: 430px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  flex-wrap: wrap;
+.buttonSubmit {
+  width: 100px;
+  height: 30px;
+  border-radius: 10px;
+  border: 2px solid black;
+  margin-top: 10px;
+  font-size: medium;
+  font-weight: 600;
+  cursor: pointer;
 }
 
-.genderWrap {
-  display: flex;
-  justify-content: center;
-}
-
-.dataSupport {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.dataSupport .labelForm {
-  font-size: 24px;
+.buttonSubmit:hover {
+  background-color: rgb(75, 159, 45);
 }
