+++ src/shared/ui/Card/index.ts
@@ -1,1 +1,1 @@
-export { Card } from './Card';
+export { Card } from './ui/Card';
