+++ src/shared/ui/Card/ui/Card.module.css
